[#18] EnumInitHelper error handling improvement
EnumInitHelper: The error suppliers returns errors instead of throwing
Tests: splited for improved readability and maintainability

Resolves #18
